refactors, renames & better organization
