+++ .github/workflows/publish.yml
@@ -12,13 +12,14 @@ jobs:
       - uses: actions/checkout@v4
       - uses: oven-sh/setup-bun@v1
 
+      # do we really need this one?
       - uses: actions/setup-node@v3
         with:
-          node-version: '20'
-          registry-url: 'https://registry.npmjs.org'
+          node-version: 18.x
 
-      - run: yarn
-      - run: yarn build
+      - run: bun install
+      - run: bun run build
+
       - uses: JS-DevTools/npm-publish@v3
         id: publish
         with:
