+++ src/index.ts
@@ -1,4 +1,3 @@
 export {UserAction, Signals, Signal} from './algorithm';
-export {FirstBatch} from './client';
+export {FirstBatch, type FirstBatchConfig} from './client';
 export {Pinecone, Weaviate, Typesense, Supabase} from './vector';
-export type {FirstBatchConfig} from './config';
