+++ src/lossy/base/index.ts
@@ -1,4 +1,4 @@
-import {CompressedVector, Vector} from '../vector/types';
+import {CompressedVector, Vector} from '../../vector/types';
 
 /** Base class for lossy compression algorithms. */
 export abstract class BaseLossy {
