+++ src/lossy/index.ts
@@ -1,4 +1,2 @@
-export * from './kmeans2';
-export * from './pqt';
 export * from './product';
 export * from './scalar';
