+++ src/lossy/product/kmeans2.ts
