+++ src/lossy/product/pqt.ts
