+++ src/vector/index.ts
@@ -1,6 +1,6 @@
 export * from './integrations';
 export * from './fetch';
 export * from './query';
-export type * from './types';
 export * from './utils';
 export * from './metadata';
+export type * from './types';
